refactor(bridge): extract profile picture FFI wrapper
